+++ src/Farmer/Builders/Builders.NatGateway.fs
@@ -20,6 +20,7 @@ type NatGatewayConfig =
                 // Currently just generate with a single public IP.
                 {
                     PublicIpAddress.Name = ResourceName $"{this.Name.Value}-publicip-1"
+                    AvailabilityZone = None
                     Location = location
                     Sku = Sku.Standard
                     AllocationMethod = AllocationMethod.Static